feat: support for algolia doc and ingestion too
