+++ scripts/package.json
@@ -12,8 +12,7 @@
     "lint:deadcode": "knip",
     "pre-commit": "node ./ci/husky/pre-commit.mjs",
     "pushGeneratedCode": "yarn runScript dist/ci/codegen/pushGeneratedCode.js",
-    "pushToAlgoliaDoc": "yarn runScript dist/ci/codegen/pushToAlgoliaDoc.js",
-    "pushToRepo": "yarn runScript dist/ci/codegen/pushToRepo.js",
+    "pushToRepository": "yarn runScript dist/ci/codegen/pushToRepository.js $*",
     "runScript": "NODE_NO_WARNINGS=1 node --enable-source-maps",
     "setRunVariables": "yarn runScript dist/ci/githubActions/setRunVariables.js",
     "spreadGeneration": "yarn runScript dist/ci/codegen/spreadGeneration.js",
